// src/app/type.d.ts
export type Tipo = "usuario" | "paciente" | "especialista" | "administrador";
export type Dia = "lunes" | "martes" | "miercoles" | "jueves" | "viernes" | "sabado" | "domingo";
export type Permiso = "usuario" | "administrador" | "especialista" | "todos" | "paciente";
export type Estado = "pendiente" | "realizado" | "cancelado" | "rechazado";